pyproject.toml
[project]
name = "django-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-amis-render>=0.0.8",
    "django-amisproxy>=0.3.0",
    "django-apscheduler>=0.7.0",
    "django-echarts>=0.6.0",
    "django-rest-admin-plus>=0.2.0",
    "drf-api-logger>=1.1.19",
    "numpy>=2.3.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pyecharts>=2.0.8",
    "python-dotenv>=1.1.1",
    "whitenoise>=6.9.0",
]
